Issues-33: Moved SimpleTranstionActivity to separate module.
