+++ app/src/main/res/values/colors.xml
@@ -6,9 +6,4 @@
 
     <color name="grey">#787878</color>
     <color name="white">#fafafa</color>
-
-    <color name="green_900">#33691E</color>
-    <color name="yellow_700">#FFAB00</color>
-    <color name="pink_700">#C51162</color>
-    <color name="brown_800">#4E342E</color>
 </resources>
