+++ app/src/main/res/values/styles.xml
@@ -18,8 +18,4 @@
         <item name="contentInsetEnd">0dp</item>
     </style>
 
-    <style name="AppTheme.TransitionAnimation" parent="AppTheme">
-        <item name="android:windowContentTransitions">true</item>
-    </style>
-
 </resources>
