+++ simpletransition/build.gradle
@@ -1,0 +1,21 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-kapt'
+apply from: '../buildsystem/dependencies.gradle'
+
+android {
+
+	compileSdkVersion versions.compileSdk
+
+	defaultConfig {
+		minSdkVersion versions.minSdk
+		targetSdkVersion versions.targetSdk
+	}
+
+}
+
+dependencies {
+	implementation libraries.kotlin
+	implementation libraries.supportAppCompat
+	implementation libraries.supportDesign
+}
