+++ simpletransition/src/main/AndroidManifest.xml
@@ -1,0 +1,18 @@
+<manifest package="com.kirchhoff.simpletransition"
+	xmlns:android="http://schemas.android.com/apk/res/android">
+
+	<application>
+
+		<activity
+			android:name=".SimpleTransitionActivity"
+			android:label="@string/simple.transition.title"
+			android:theme="@style/AppTheme.TransitionAnimation" />
+
+		<activity
+			android:name=".SimpleTransitionDetailsActivity"
+			android:label="@string/simple.transition.details.title"
+			android:theme="@style/AppTheme.TransitionAnimation" />
+
+	</application>
+
+</manifest>
