+++ simpletransition/src/main/res/layout/a_simple_transition.xml
