+++ simpletransition/src/main/res/layout/a_simple_transition_details.xml
