+++ simpletransition/src/main/res/layout/r_simple_transition.xml
