+++ simpletransition/src/main/res/values/colors.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+	<color name="green_900">#33691E</color>
+	<color name="yellow_700">#FFAB00</color>
+	<color name="pink_700">#C51162</color>
+	<color name="brown_800">#4E342E</color>
+</resources>
